requirements.txt
feedparser>=6.0
beautifulsoup4>=4.12
twikit>=2.3
python-dotenv>=1.0
requests>=2.28
kokoro-onnx>=0.4.7
lameenc>=1.8
soundfile>=0.13
